move dataset creation in dataset.py
